requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
Flask-SQLAlchemy==3.1.1
colorlog==6.8.0
Flask-Login==0.6.3
